feat: Handle updates and deletions of reviews using post middleware
Implemented a post middleware in `reviewSchema` to handle updates and deletions of reviews. The middleware is triggered for operations matching /^findOneAnd/, ensuring it runs after operations like findOneAndUpdate and findOneAndDelete.

When a review document (`reviewDoc`) is found or updated, it calls `calcAverageRatings()` on the model constructor, passing the tour ID from `reviewDoc.tour`. This ensures that average ratings for associated tours are dynamically updated after each review operation.

This approach enhances the application's efficiency by automating the recalculation of average ratings without manual intervention, improving data accuracy and user experience.